Support optional INIT_API_KEY for token
Adds support for an optional INIT_API_KEY env variable that if set
will require the same value be provided when requesting token.

- `GET /scim/token` will requre HEADER `INIT_API_KEY` with
the value of the env variable `INIT_API_KEY` if it is set.

This is a special use case where a devt environment is publicly
available (eg Azure hosted) but needs a level of restriction.